+++ UI/frontend-plugins/frontend-tools/output-timer.cpp
@@ -283,18 +283,18 @@ static void SaveOutputTimer(obs_data_t *save_data, bool saving, void *)
 			obj = obs_data_create();
 
 		ot->ui->streamingTimerHours->setValue(
-			obs_data_get_int(obj, "streamTimerHours"));
+			(int)obs_data_get_int(obj, "streamTimerHours"));
 		ot->ui->streamingTimerMinutes->setValue(
-			obs_data_get_int(obj, "streamTimerMinutes"));
+                                                (int)obs_data_get_int(obj, "streamTimerMinutes"));
 		ot->ui->streamingTimerSeconds->setValue(
-			obs_data_get_int(obj, "streamTimerSeconds"));
+                                                (int)obs_data_get_int(obj, "streamTimerSeconds"));
 
 		ot->ui->recordingTimerHours->setValue(
-			obs_data_get_int(obj, "recordTimerHours"));
+                                              (int)obs_data_get_int(obj, "recordTimerHours"));
 		ot->ui->recordingTimerMinutes->setValue(
-			obs_data_get_int(obj, "recordTimerMinutes"));
+                                                (int)obs_data_get_int(obj, "recordTimerMinutes"));
 		ot->ui->recordingTimerSeconds->setValue(
-			obs_data_get_int(obj, "recordTimerSeconds"));
+                                                (int)obs_data_get_int(obj, "recordTimerSeconds"));
 
 		ot->ui->autoStartStreamTimer->setChecked(
 			obs_data_get_bool(obj, "autoStartStreamTimer"));
